backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
# passlib 1.7.4 calls bcrypt.__about__ which was removed in bcrypt 4.1+.
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
openpyxl==3.1.5
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
httpx==0.27.2
